// sistema/src/components/shared/Interfaces.ts
export interface ImagenState {
  archivo: File | null
  archivoName: string
}
// ...
export interface productosValuesModificate {
  nombre: string
}